+++ data/bob.txt
@@ -1,1 +1,2 @@
 [D][ ] return book (by: Dec 02 2019 18:00)
+[D][ ] return book (by: Dec 02 2019 18:00)
