+++ src/main/java/Bob/BobException.java
@@ -1,0 +1,8 @@
+package bob;
+
+class BobException extends Exception {
+
+    public BobException(String message) {
+        super(message);
+    }
+}
