+++ src/test/java/bob/task/TaskTest.java
@@ -1,0 +1,15 @@
+package bob.task;
+
+public class TaskTest {
+    public void TaskConstructorTest() {
+        Task task = new Task("description");
+        assertEquals(task.getDescription(), "description");
+        assertEquals(task.getStatusIcon(), " ");
+    }
+
+    public void markAsDoneTest() {
+        Task task = new Task("description");
+        task.markAsDone();
+        assertEquals(task.getStatusIcon(), "X");
+    }
+}
